chore: SocialImageListView에 있던 디테일 이미지 관련 파일 분리
